pyproject.toml
[tool.poetry]
name = "typespace"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.0"
fastapi = {extras = ["all"], version = "^0.115.4"}
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
pyjwt = "^2.9.0"
aiofiles = "^24.1.0"
authlib = "^1.3.2"
alembic = "^1.14.0"
black = {extras = ["d"], version = "^24.10.0"}
aiosmtplib = "^3.0.2"
bcrypt = "^4.2.1"
celery = {extras = ["redis"], version = "^5.4.0"}
asgiref = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
